est: renamed client and server to estclient and estserver
